requirement.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
fonttools==4.56.0
frozendict==2.4.6
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
moviepy==2.1.2
multitasking==0.0.11
nltk==3.9.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.7
proglog==0.1.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soupsieve==2.6
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
yfinance==0.2.55
